+++ src/org/schema/constants/traits/Credential.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace org\schema\constants\traits;
+
+/**
+ * The Credential properties enumeration.
+ */
+trait Credential
+{
+    const string CREDENTIAL_CATEGORY = 'credentialCategory' ;
+    const string RECOGNIZED_BY       = 'recognizedBy' ;
+    const string VALID_FOR           = 'validFor' ;
+    const string VALID_IN            = 'validIn' ;
+}