+++ src/org/schema/constants/traits/EducationalOccupationCredential.php
@@ -4,10 +4,8 @@ namespace org\schema\constants\traits;
 
 trait EducationalOccupationCredential
 {
+    use Credential ;
+
     const string COMPETENCY_REQUIRED = 'competencyRequired' ;
-    const string CREDENTIAL_CATEGORY = 'credentialCategory' ;
-    const string RECOGNIZED_BY       = 'recognizedBy' ;
-    const string VALID_FOR           = 'validFor' ;
-    const string VALID_IN            = 'validIn' ;
     const string EDUCATIONAL_LEVEL   = 'educationalLevel' ;
 }