+++ src/org/schema/constants/traits/Properties.php
@@ -22,6 +22,7 @@ trait Properties
         ConstraintNode ,
         ContactPoint ,
         CreativeWork ,
+        Credential ,
         DataFeedItem ,
         Dataset ,
         DefinedRegion ,